# Remove dead code in SavedVariable