Stack from [ghstack](https://github.com/ezyang/ghstack):
* **#59838 Remove dead code in SavedVariable**
* #59837 Avoid un-necessary unwrapping of Tensor in SavedVariable

Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Differential Revision: [D29069214](https://our.internmc.facebook.com/intern/diff/D29069214)