@Varal7 2021-06-11
> code removal!
> 
> It would be nice to do a small benchmark of the impact of this stack on perf.
> In particular, using the Timer class with instruction count. We can chat about this offline if you want.

Thanks! As a spoiler, I already ran the following test:
```
# Benchmark from: https://github.com/pytorch/pytorch/pull/56017#issuecomment-853388987

import torch
import time

x = torch.zeros(1, device='cpu', requires_grad=True)
for _ in range(100000):
    x = x * x

start = time.time()
x.backward()
end = time.time()
print(end-start)

# Without the patch:

➜ for i in `seq 1 10`; do python bench.py; done
0.9253489971160889
0.9564845561981201
0.9363410472869873
0.897282600402832
0.8867707252502441
0.8611898422241211
0.8699934482574463
0.8362593650817871
0.8579938411712646
0.8739769458770752

# With the patch:

➜ for i in `seq 1 10`; do python bench.py; done
0.5738115310668945
0.5449838638305664
0.5333981513977051
0.5343718528747559
0.5725717544555664
0.5140812397003174
0.5454022884368896
0.5519816875457764
0.531273365020752
0.5399165153503418

```