@Varal7 2021-06-14
Benchmark!

## Wall time

```python
# bench.py
x = torch.zeros(1, device='cpu', requires_grad=True)

start_fw = time.time()
for _ in range(100000):
    x = x * x
end_fw = time.time()

start_bw = time.time()
x.backward()
end_bw = time.time()
print("Forward\t", end_fw-start_fw, end="\t")
print("Backward\t", end_bw-start_bw)
```

Without the patch:

```
➜ for i in `seq 1 10`; do python bench.py; done
Forward  0.47095823287963867    Backward         0.9542675018310547
Forward  0.388355016708374      Backward         0.9605531692504883
Forward  0.42574310302734375    Backward         1.0612819194793701
Forward  0.38661670684814453    Backward         0.9773132801055908
Forward  0.37992262840270996    Backward         0.9450616836547852
Forward  0.3741035461425781     Backward         0.8979041576385498
Forward  0.423450231552124      Backward         1.049081802368164
Forward  0.39497900009155273    Backward         0.9403750896453857
Forward  0.3828461170196533     Backward         0.9505488872528076
Forward  0.36185240745544434    Backward         0.9551246166229248
```

With the patch:

```

➜ for i in `seq 1 10`; do python bench.py; done
Forward  0.33322834968566895    Backward         0.6979451179504395
Forward  0.34907960891723633    Backward         0.7061493396759033
Forward  0.34142494201660156    Backward         0.6910700798034668
Forward  0.34435534477233887    Backward         0.6630330085754395
Forward  0.3300626277923584     Backward         0.701082706451416
Forward  0.4510471820831299     Backward         0.874150276184082
Forward  0.3268928527832031     Backward         0.7102131843566895
Forward  0.3682827949523926     Backward         0.6823954582214355
Forward  0.3338043689727783     Backward         0.6908135414123535
Forward  0.3473036289215088     Backward         0.7074475288391113
```

## Instruction count

100 runs per measurement, 1 thread

### Forward

```
stmt: y = (x * x).sum()
setup: x = torch.randn(5, requires_grad=True)
```

Without the patch:

```
                           All          Noisy symbols removed
    Instructions:      2507869                    2492089
    Baseline:              899                        801
```

With the patch:

```
                           All          Noisy symbols removed
    Instructions:      2242869                    2214453
    Baseline:              899                        801
```


### Backward

```
stmt: y.backward(retain_graph=True)
setup: x = torch.randn(5, requires_grad=True); y = (x * x).sum()
```

Without the patch:
```
                           All          Noisy symbols removed
    Instructions:      7286245                    7178227
    Baseline:              919                        801
```

With the patch:

```
                           All          Noisy symbols removed
    Instructions:      6939964                    6847261
    Baseline:              899                        801
```